transfer compute_edge_lengths to lib
